Fix CSV write
