+++ awswrangler/distributed/ray/modin/_utils.py
@@ -1,5 +1,6 @@
 """Modin on Ray utilities (PRIVATE)."""
-from typing import Any, Callable, Dict, List, Optional, Union
+from dataclasses import dataclass
+from typing import Any, Callable, Dict, List, Optional, Set, Union
 
 import modin.pandas as modin_pd
 import pandas as pd
@@ -8,6 +9,7 @@ from modin.distributed.dataframe.pandas import from_partitions
 from ray.data._internal.arrow_block import ArrowBlockAccessor, ArrowRow
 from ray.data._internal.remote_fn import cached_remote_fn
 
+from awswrangler import exceptions
 from awswrangler._arrow import _table_to_df
 
 
@@ -43,3 +45,24 @@ def _to_modin(
 
 def _arrow_refs_to_df(arrow_refs: List[Callable[..., Any]], kwargs: Optional[Dict[str, Any]]) -> modin_pd.DataFrame:
     return _to_modin(dataset=ray.data.from_arrow_refs(arrow_refs), to_pandas_kwargs=kwargs)
+
+
+@dataclass
+class ParamConfig:
+    default: Any
+    supported_values: Optional[Set[Any]] = None
+
+
+def check_parameters(pandas_kwargs: Dict[str, Any], supported_params: Dict[str, ParamConfig]) -> None:
+    for pandas_arg_key, pandas_args_value in pandas_kwargs.items():
+        if pandas_arg_key not in supported_params:
+            raise exceptions.InvalidArgument(f"Unsupported Pandas parameter for PyArrow loader: {pandas_arg_key}")
+
+        param_config = supported_params[pandas_arg_key]
+        if param_config.supported_values is None:
+            continue
+
+        if pandas_args_value not in param_config.supported_values:
+            raise exceptions.InvalidArgument(
+                f"Unsupported Pandas parameter value for PyArrow loader: {pandas_arg_key}={pandas_args_value}",
+            )
