+++ tests/load/test_s3.py
@@ -127,7 +127,7 @@ def test_s3_delete_objects(path, path2, benchmark_time):
     assert len(wr.s3.list_objects(f"{path2}delete-test*")) == 0
 
 
-@pytest.mark.parametrize("benchmark_time", [240])
+@pytest.mark.parametrize("benchmark_time", [20])
 def test_s3_read_csv_simple(benchmark_time):
     path = "s3://nyc-tlc/csv_backup/yellow_tripdata_2021-0*.csv"
     with ExecutionTimer("elapsed time of wr.s3.read_csv() simple") as timer:
